requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.3.0
boto3==1.40.27
botocore==1.40.27
certifi==2025.8.3
cffi
et_xmlfile==2.0.0
h11==0.16.0
idna==3.10
jmespath==1.0.1
minio==7.2.16
numpy==2.3.2
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.3.2
pycparser==2.22
pycryptodome==3.23.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
s3transfer==0.14.0
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
selenium 
webdriver-manager